% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySqrt.R
\name{mySqrt}
\alias{mySqrt}
\title{Compute square root of a number}
\usage{
mySqrt(x)
}
\arguments{
\item{x}{a postive number}
}
\value{
The square root of \code{x}. If x is less than zero then return #' the compelx square root of x
See \url{https://en.wikipedia.org/wiki/Square_root}
for more information on how this is all supposed to work.
}
\description{
\code{mySqrt} returns the square root of a number x
}
\examples{
mySqrt(x=4)
}
\author{
Steve Pittard \href{mailto:wsp@emory.edu}{wsp@emory.edu}
}
