% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfind.R
\docType{package}
\name{rootfind}
\alias{rootfind}
\title{rootfind: a package for finding roots}
\description{
rootfind implements a single function called mySqrt
It has three main goals:
}
\details{
\itemize{
\item Identify important devtools concepts
\item Provide examples of roxygen
\item Provide a package container for the mySqrt function
}
}
